+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Upcoming
 
+- Add `disabledProviders` config that tracks the list of disabled linter providers
 - Enhance linter provider error message to include opening instructions for dev tools
 
 ## 2.1.2